schemachange/requirements.txt
snowflake-connector-python==2.8.1
schemachange==3.5.3
j2cli==0.3.10
